requirements.txt
Flask==3.1.3
Werkzeug==3.1.6
python-dotenv==1.2.1

Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.43
psycopg[binary]==3.2.12

Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==4.1.1
email-validator==2.3.0

cryptography==46.0.5
pyotp==2.9.0
bleach==6.2.0
argon2-cffi==23.1.0

redis==7.2.0
gunicorn==25.1.0
uvicorn==0.36.0
fastapi==0.118.2

rich==14.3.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
playwright==1.58.0

pytest==8.4.2
pytest-mock==3.15.1
